THORN-586: Auto-calculate nof_splits, allow for setting nbuffers

+++ src/ska_low_mccs_daq/cdaq/TensorCorrelatorData.h
@@ -45,7 +45,7 @@ public:
                           uint32_t nof_samples,
                           uint8_t nof_pols,
                           uint16_t nof_active_antennas,
-                          uint8_t nof_splits = 1,
+                          uint16_t nof_splits = 1,
                           uint8_t nbuffers = 4);
     ~TensorCrossCorrelator() override;
     TccDoubleBuffer *double_buffer;
@@ -64,7 +64,7 @@ private:
     cu::Stream stream_;
 
     // nof_splits_ must precede samplesExt_ (init-list order = declaration order)
-    uint8_t nof_splits_;
+    uint16_t nof_splits_;
 
     // Extents next (sizes used by buffers)
     multi_array::extent<5> samplesExt_;
@@ -80,7 +80,7 @@ private:
 
     // Streaming helpers
     void copy_tail(const uint8_t *host_base, size_t split_start, size_t from, size_t to);
-    void try_stream_partial(uint8_t &split, size_t &split_streamed);
+    void try_stream_partial(uint16_t &split, size_t &split_streamed);
 
     // config
     uint32_t nof_channels;
